Reorganize code to prevent cyclic imports
