+++ ui/src/components/forms/stop-area/index.ts
@@ -1,3 +1,4 @@
 export * from './SelectMemberStopsDropdown';
 export * from './ConfirmStopAreaDeletionDialog';
 export * from './stopAreaFormSchema';
+export * from './useUpsertStopArea';
