+++ ui/src/hooks/stop-registry/stop-areas/index.ts
@@ -1,3 +1,2 @@
 export * from './useDeleteStopArea';
 export * from './useGetStopAreaById';
-export * from './useUpsertStopArea';
