+++ ui/src/types/StopSearchConditions.ts
@@ -1,0 +1,6 @@
+export type StopSearchConditions = {
+  elyNumber: string;
+  municipalities: string;
+  address?: string;
+  labelOrName?: string;
+};
