+++ ui/src/types/stopAreaByIdResult.ts
@@ -1,0 +1,8 @@
+import {
+  StopAreaFormFieldsFragment,
+  StopAreaMemberFieldsFragment,
+} from '../generated/graphql';
+
+export type StopAreaByIdResult = Omit<StopAreaFormFieldsFragment, 'members'> & {
+  members: Array<StopAreaMemberFieldsFragment> | undefined;
+};
