+++ ui/src/utils/stop-registry/index.ts
@@ -1,2 +1,3 @@
 export * from './stopPlace';
 export * from './alternativeNames';
+export * from './buildSearchStopsGqlQueryVariables';
